# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/balance/{id}": {
      "get": {
        "tags": ["Balances"],
        "summary": "Get balance by ID",
        "description": "Returns a single balance",
        "operationId": "getBalanceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of balance to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "balance found"
          },
          "404": {
            "description": "balance not found"
          }
        }
      },
      "put": {
        "tags": ["Balances"],
        "summary": "Update balance by ID",
        "description": "Updates a single user",
        "operationId": "updateBalanceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of balance to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Balance",
            "in": "body",
            "description": "balance object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "cash": {
                  "type": "number",
                  "default": 10000
                },
                "users_id": {
                  "type": "number",
                  "default": 6
                }
              }
            }
          }
        ],

        "responses": {
          "200": {
            "description": "balance updated"
          },
          "404": {
            "description": "balance not found"
          }
        }
      },

      "delete": {
        "tags": ["Balances"],
        "summary": "delete balance by ID",
        "description": "delete user balance",
        "operationId": "deleteBalances",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of balance to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "balance delted"
          },
          "404": {
            "description": "balance not found"
          }
        }
      }
    },
...
  }
}